\name{mice_default}
\alias{mice_default}
\title{Find the default methods given by \pkg{mice}}
\usage{
mice_default(vec, dat)
}
\arguments{
  \item{vec}{a vector of the variable classes.}

  \item{dat}{the data corresponding to vec.}
}
\value{
  A vector of the default imputation method generated by
  mice.
}
\description{
  Find the default methods given by \pkg{mice}
}
\author{
  Xiaoyue Cheng <\email{xycheng@iastate.edu}>
}
