% Please edit documentation in R/MissingDataGUI-package.r
\docType{package}
\name{MissingDataGUI-package}
\alias{MissingDataGUI-package}
\title{A Graphical User Interface for Exploring Missing Values in Data}
\description{
This package was designed mainly for the exploration of
missing values structure, and results of imputation, using static
graphics and numerical summaries. A graphical user interface (GUI)
makes it accessible to novice users.
}
\examples{
if (interactive()) {
    MissingDataGUI()
}

}
\references{
Xiaoyue Cheng, Dianne Cook, Heike Hofmann (2015). Visually Exploring Missing Values in Multivariable Data Using a Graphical User Interface. Journal of Statistical Software, 68(6), 1-23. doi:10.18637/jss.v068.i06
}
