% Please edit documentation in R/utils.r
\name{scale_colour_discrete}
\alias{scale_colour_discrete}
\title{Change the discrete color scale for the plots generated by ggplot2}
\usage{
scale_colour_discrete(...)
}
\arguments{
\item{...}{parameters passed into the function}
}
\description{
Change the discrete color scale for the plots generated by ggplot2
}
