% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scale_fill_discrete}
\alias{scale_fill_discrete}
\title{Change the discrete fill scale for the plots generated by ggplot2}
\usage{
scale_fill_discrete(...)
}
\arguments{
\item{...}{parameters passed into the function}
}
\description{
Change the discrete fill scale for the plots generated by ggplot2
}

