\name{winprint}
\alias{winprint}
\title{Print ggpair object in GTK+ for Windows platform}
\usage{
winprint(x, ...)
}
\arguments{
  \item{x}{ggpair object to be plotted}

  \item{...}{not used}
}
\value{
  NULL
}
\description{
  See \code{print.ggpairs} from \pkg{GGally}.
}
\author{
  Barret Schloerke <\email{schloerke@gmail.com}>, Xiaoyue
  Cheng <\email{xycheng@iastate.edu}>
}
