\docType{package}
\name{MissingDataGUI-package}
\alias{MissingDataGUI-package}
\title{A Graphical User Interface for Exploring Missing Values in Data}
\description{
  This package was designed mainly for the exploration of
  missing values structure, and results of imputation,
  using static graphics and numerical summaries. A
  graphical user interface (GUI) makes it accessible to
  novice users.
}
\examples{
if (interactive()) {
    MissingDataGUI()
}
}
