% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scale_colour_discrete}
\alias{scale_colour_discrete}
\title{Change the discrete color scale for the plots generated by ggplot2}
\usage{
scale_colour_discrete(...)
}
\arguments{
\item{...}{parameters passed into the function}
}
\description{
Change the discrete color scale for the plots generated by ggplot2
}

